requirements.txt
asgiref==3.7.2
cachetools==5.5.0
celery==5.4.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.1
django-debug-toolbar==4.1.0
django-prometheus==2.3.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
drf-yasg==1.21.5
exceptiongroup==1.1.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
model-bakery==1.11.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pycparser==2.21
PyJWT==2.7.0
prometheus_client==0.21.1
psycopg2==2.9.10
pytest==7.3.1
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.4.0
